requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
arrow==1.3.0
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
awscli==1.41.3
backoff==2.2.1
billiard==4.2.1
binaryornot==0.4.4
boto3==1.39.1
botocore==1.39.3
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.2
cookiecutter==2.6.0
cryptography==45.0.4
cycler==0.12.1
dacite==1.6.0
dagshub==0.5.10
dagshub-annotation-converter==0.1.10
dataclasses-json==0.6.7
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docutils==0.19
dpath==2.2.0
dulwich==0.23.1
dvc==3.60.1
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
faiss-cpu==1.11.0
fastapi==0.115.14
filelock==3.18.0
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
gql==3.5.3
grandalf==0.8
graphql-core==3.2.6
gto==1.7.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
iterative-telemetry==0.0.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.4
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.1
omegaconf==2.3.0
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pathvalidate==3.3.1
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.0
Pygments==2.19.2
pygtrie==2.5.0
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pywin32==310
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.7.2
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.16.0
scmrepo==3.3.11
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqltrie==0.11.2
-e git+https://github.com/Mahi-Anol/Image_search_engine_production.git@8aaffb4b6e8aeef7bc2e74e9df772d5e9eb7c81f#egg=src
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
treelib==1.8.0
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
yarl==1.20.1
zc.lockfile==3.0.post1
